fixes nuke ops bug aswell as no paradox clone

+++ monkestation/code/modules/storytellers/converted_events/solo/ghosts/nuclear_operative_ghost.dm
@@ -43,15 +43,19 @@
 	var/static/datum/team/nuclear/nuke_team
 	var/set_leader = FALSE
 	var/required_role = ROLE_NUCLEAR_OPERATIVE
+	var/datum/mind/most_experienced
 
 /datum/round_event/antagonist/solo/ghost/nuclear_operative/add_datum_to_mind(datum/mind/antag_mind)
+	if(most_experienced == antag_mind)
+		return
 	var/mob/living/current_mob = antag_mind.current
 	var/list/items = current_mob.get_equipped_items(TRUE)
 	current_mob.unequip_everything()
 	for(var/obj/item/item as anything in items)
 		qdel(item)
 
-	var/datum/mind/most_experienced = get_most_experienced(setup_minds, required_role)
+	if(!most_experienced)
+		most_experienced = get_most_experienced(setup_minds, required_role)
 	antag_mind.set_assigned_role(SSjob.GetJobType(/datum/job/nuclear_operative))
 	antag_mind.special_role = ROLE_NUCLEAR_OPERATIVE
